bandmap: only delete spots once every 10 bandmap requests

+++ bandmap.php
@@ -161,7 +161,9 @@ $json_a = array();
 
 if (intval(phpversion())>=5) {
   mysqli_query($con, "set timezone = '+00:00'");
-  mysqli_query($con, "delete from spots where time < (UTC_TIMESTAMP() - INTERVAL 60 MINUTE) or time > (UTC_TIMESTAMP() + INTERVAL 30 MINUTE);");
+  if (rand(10) < 1) {
+     mysqli_query($con, "delete from spots where time < (UTC_TIMESTAMP() - INTERVAL 60 MINUTE) or time > (UTC_TIMESTAMP() + INTERVAL 30 MINUTE);");
+  }
     # Delete spots older than 60 minutes, or spots that were made (over 30 minutes) in the future
   } 
 else {
